@@ -1,4 +1,4 @@
-class CreatePatient < ActiveRecord::Migration
+class CreatePatient < ActiveRecord::Migration[4.2]
   def change
     create_table :patients do |t|
       t.string  :name
