@@ -1,4 +1,4 @@
-class CreateCases < ActiveRecord::Migration
+class CreateCases < ActiveRecord::Migration[4.2]
   def change
     create_table :cases do |t|
       t.integer :patient_id
