@@ -1,4 +1,4 @@
-class AddTimestampsToPatientsAndCases < ActiveRecord::Migration
+class AddTimestampsToPatientsAndCases < ActiveRecord::Migration[4.2]
   def change
     add_column :patients, :created_at, :datetime
     add_column :patients, :updated_at, :datetime
