@@ -1,4 +1,4 @@
-class CreateProcedures < ActiveRecord::Migration
+class CreateProcedures < ActiveRecord::Migration[4.2]
   def change
     create_table :procedures do |t|
       t.integer :case_id, null: false
