@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -16,85 +15,79 @@ ActiveRecord::Schema.define(version: 20160327101859) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "appointments", force: :cascade do |t|
-    t.integer  "patient_id"
-    t.integer  "trip_id"
-    t.integer  "order"
-    t.date     "date"
+  create_table "appointments", id: :serial, force: :cascade do |t|
+    t.integer "patient_id"
+    t.integer "trip_id"
+    t.integer "order"
+    t.date "date"
     t.datetime "start"
     t.datetime "end"
-    t.string   "location"
+    t.string "location"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.index ["patient_id"], name: "index_appointments_on_patient_id"
+    t.index ["trip_id", "start", "location"], name: "index_appointments_on_trip_id_and_start_and_location"
   end
 
-  add_index "appointments", ["patient_id"], name: "index_appointments_on_patient_id", using: :btree
-  add_index "appointments", ["trip_id", "start", "location"], name: "index_appointments_on_trip_id_and_start_and_location", using: :btree
-
-  create_table "attachments", force: :cascade do |t|
-    t.integer  "record_id"
-    t.string   "record_type"
-    t.text     "description"
-    t.string   "document_file_name"
-    t.string   "document_content_type"
-    t.integer  "document_file_size"
+  create_table "attachments", id: :serial, force: :cascade do |t|
+    t.integer "record_id"
+    t.string "record_type"
+    t.text "description"
+    t.string "document_file_name"
+    t.string "document_content_type"
+    t.integer "document_file_size"
     t.datetime "document_updated_at"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["record_id"], name: "index_attachments_on_record_id"
+    t.index ["record_type", "record_id"], name: "index_attachments_on_record_type_and_record_id"
   end
 
-  add_index "attachments", ["record_id"], name: "index_attachments_on_record_id", using: :btree
-  add_index "attachments", ["record_type", "record_id"], name: "index_attachments_on_record_type_and_record_id", using: :btree
-
-  create_table "cases", force: :cascade do |t|
-    t.integer  "patient_id"
-    t.string   "anatomy"
-    t.string   "side"
-    t.string   "status",     default: "active"
+  create_table "cases", id: :serial, force: :cascade do |t|
+    t.integer "patient_id"
+    t.string "anatomy"
+    t.string "side"
+    t.string "status", default: "active"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.index ["status"], name: "index_cases_on_status"
   end
 
-  add_index "cases", ["status"], name: "index_cases_on_status", using: :btree
-
-  create_table "missions", force: :cascade do |t|
-    t.string "name",                 null: false
-    t.string "country",    limit: 2
+  create_table "missions", id: :serial, force: :cascade do |t|
+    t.string "name", null: false
+    t.string "country", limit: 2
     t.string "location"
     t.string "facility"
-    t.date   "begin_date"
-    t.date   "end_date"
+    t.date "begin_date"
+    t.date "end_date"
+    t.index ["begin_date"], name: "index_missions_on_begin_date"
   end
 
-  add_index "missions", ["begin_date"], name: "index_missions_on_begin_date", using: :btree
-
-  create_table "missions_teams", force: :cascade do |t|
+  create_table "missions_teams", id: :serial, force: :cascade do |t|
     t.integer "mission_id", null: false
-    t.integer "team_id",    null: false
+    t.integer "team_id", null: false
+    t.index ["team_id"], name: "index_missions_teams_on_team_id"
   end
 
-  add_index "missions_teams", ["team_id"], name: "index_missions_teams_on_team_id", using: :btree
-
-  create_table "patients", force: :cascade do |t|
-    t.string   "name"
-    t.date     "birth"
-    t.string   "gender",     limit: 10
-    t.date     "death"
-    t.string   "status",                default: "active"
+  create_table "patients", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.date "birth"
+    t.string "gender", limit: 10
+    t.date "death"
+    t.string "status", default: "active"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.index ["name", "birth"], name: "index_patients_on_name_and_birth"
+    t.index ["status"], name: "index_patients_on_status"
   end
 
-  add_index "patients", ["name", "birth"], name: "index_patients_on_name_and_birth", using: :btree
-  add_index "patients", ["status"], name: "index_patients_on_status", using: :btree
-
-  create_table "procedures", force: :cascade do |t|
-    t.integer "case_id",        null: false
+  create_table "procedures", id: :serial, force: :cascade do |t|
+    t.integer "case_id", null: false
     t.integer "appointment_id"
-    t.jsonb   "data"
+    t.jsonb "data"
   end
 
-  create_table "teams", force: :cascade do |t|
+  create_table "teams", id: :serial, force: :cascade do |t|
     t.string "name", null: false
   end
 
